https://sonarcloud.io/project/issues?branch=dev&id=blish-hud_Blish-HUD&issues=AX3bN1vzlgoyO_aM-INx&open=AX3bN1vzlgoyO_aM-INx https://sonarcloud.io/project/issues?branch=dev&id=blish-hud_Blish-HUD&issues=AX3bN1vzlgoyO_aM-INy&open=AX3bN1vzlgoyO_aM-INy

+++ Blish HUD/Controls/TrackBar.cs
@@ -109,7 +109,7 @@ namespace Blish_HUD.Controls {
         private void MinMaxChanged() {
             tenIncrements.Clear();
             for (int i = 0; i < 11; i++) {
-                tenIncrements.Add((float)(this.MaxValue - this.MinValue) * 0.1f * i + (float)this.MinValue);
+                tenIncrements.Add((this.MaxValue - this.MinValue) * 0.1f * i + this.MinValue);
             }
         }
